refactor(chat): remove chat streaming functionality and update related components
